deleting files with date errors + broken urls
BELDING TANK TECHNOLOGIES, INC.:  SAR was dated Dec. 19, 2024. It was changed to Dec. 19, 2023.

RIVERSIDE - MOSSBACK CPF: SAR was dated Nov. 9, 2024. It was changed to Nov. 9, 2023.

HOLCIM (US) INC. DBA LAFARGE ALPENA PLANT: SAR was dated Nov. 17, 2024. Could not find equivalent SAR from the year before, so changed it to 2023 manually and left the record in the dataset. Link is still broken, but interested party could ask EGLE if there was an SAR from Nov. 17, 2023 that was removed.